Adicionado separação entre os domínios e menos um botão
