--- domain_trader/lib/View/home_page.dart
@@ -1,78 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HomePage extends StatefulWidget {
-  const HomePage({super.key});
-
-  @override
-  State<HomePage> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: Column(
-          children: [
-            Expanded(
-              child: Container(
-                color: Colors.white,
-                child: Padding(
-                  padding: const EdgeInsets.only(left: 20, right: 20, top: 12, bottom: 12),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      const Text(
-                        'Domain Trader',
-                        style: TextStyle(
-                          fontSize: 24,
-                          fontWeight: FontWeight.bold
-                        ),
-                      ),
-                      OutlinedButton(
-                        onPressed: () {},
-                        child: const Icon(Icons.person),
-                      )
-                    ],
-                  ),
-                ),
-              )
-            ),
-            Expanded(
-              flex: 8,
-              child: Container(
-                color: Colors.grey,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: [
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        Container(
-                          margin: const EdgeInsets.all(20),
-                          width: 410,
-                          height: 100,
-                          color: Colors.blueGrey,
-                          child: const Padding(
-                            padding: EdgeInsets.all(18.0),
-                            child: Row(
-                              mainAxisAlignment: MainAxisAlignment.start,
-                              children: [
-                                Text('Valor'),
-                              ],
-                            )
-                          ),
-                        )
-                      ],
-                    ),
-                  ],
-                ),
-              )
-            ),
-          ],
-        )
-      ),
-    );
-  }
-}
